requirements.txt
# ============================================================================
# COUNTERPARTY RECOMMENDATION ENGINE - DEPENDENCIES
# ============================================================================

# INSTALLATION INSTRUCTIONS:
# 1. Install Python dependencies:
#    pip install -r requirements.txt
#
# 2. Download spaCy language model:
#    python -m spacy download en_core_web_sm
#
# 3. Install Node.js dependencies (for frontend):
#    cd counterparty-ui
#    npm install
#
# 4. Set up environment variables:
#    cp .env.example .env
#    # Edit .env and add your GEMINI_API_KEY
#
# 5. Run backend:
#    uvicorn app.main:app --reload
#
# 6. Run frontend (in separate terminal):
#    cd counterparty-ui
#    npm start

# ============================================================================
# PYTHON BACKEND DEPENDENCIES
# ============================================================================

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database & Validation
sqlalchemy==2.0.35
pydantic==2.9.0
pydantic-settings==2.5.2

# NLP & Text Processing
spacy==3.7.2
# After installing spacy, run: python -m spacy download en_core_web_sm

# Document Processing
pdfminer.six==20221105
python-docx==1.1.0

# Export & Templating
reportlab==4.0.9
jinja2==3.1.3

# HTTP Client
httpx==0.27.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
hypothesis==6.112.0

# ============================================================================
# FRONTEND DEPENDENCIES (Angular)
# ============================================================================
# These are managed by npm/package.json in the counterparty-ui directory
# Run: cd counterparty-ui && npm install
#
# Key dependencies:
# - @angular/core: ^18.0.0
# - @angular/material: ^18.0.0
# - @angular/cdk: ^18.0.0
# - rxjs: ~7.8.0
# - typescript: ~5.4.0
#
# See counterparty-ui/package.json for complete list
